:truck: rename: my 위젯 폴더 구조 정리
