+++ src/widget/my/ui/main/my/Stats.tsx
